Ensure a consistent, language-independent ID for the NewsBox portlet
Previously the NewsBox portlet ID would vary depending on the (user) language and the contents of [[MediaWiki:Newsbox-title]].
For example, for English (where the default value for that msg is "ShoutWiki Messages") the ID would be often something like p-ShoutWiki_messages, and for Finnish where the msg value would be "ShoutWikin viestit" the ID would likewise be p-ShoutWikin_viestit etc.
This meant that targeting the NewsBox via CSS was outright impossible and that customizing the [[MediaWiki:Newsbox-title]] message would result in the NewsBox portlet ID changing as well.

This is now fixed and now the NewsBox portlet has the consistent ID "p-newsbox-title", derived from the message key instead of contents.

Thanks to Jon Robson on T270447 for pointing me to the right direction!

References T270447

Change-Id: I5f31ea7bd0e9fcb82f2527ebb69daff6769068fb