Make NewsBox relevant again
* Major code cleanup -- renderForMonaco()'s guts merged into getNewsBoxHTML() (which used to be render)
* Uses the SkinAfterPortlet hook to support MonoBook & Vector in MW 1.39+ as well as various other skins
** Note, however, that many skins need to be individually patched... see https://gerrit.wikimedia.org/r/q/topic:make-NewsBox-relevant-again for the skin patches
* Nimbus calls getNewsBoxHTML() directly in its main PHP template class and thus sorta side-steps T216851 and one could even argue that "fixes" the issue

What was done to the SkinBuildSidebar hook in core between 1.35 and 1.39 is not just bad, it is simply unforgivable.
It was an elegant, works-in-most-skins solution, as opposed to this, which is a pile of crude hacks on top of other crude hacks since there currently seems to be no way to render a "portlet" the way YOU, as an extension developer, wish to do it.
Raw HTML is bad only if you allow users to enter it, it is NOT an anti-pattern to let DEVELOPERS use raw HTML if and when they need it.

Anyway, NewsBox now mostly works in many skins again. I'm hardly *happy* with the code quality, but it is what it is.

Change-Id: Idd683ce878409d743ed50d44de651f83764ce41e